+++ packages/async/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.2](https://github.com/rambler-digital-solutions/rambler-common/compare/@rambler-tech/async@0.1.1...@rambler-tech/async@0.1.2) (2024-09-27)
+
+**Note:** Version bump only for package @rambler-tech/async
+
+## [0.1.1](https://github.com/rambler-digital-solutions/rambler-common/compare/@rambler-tech/async@0.1.0...@rambler-tech/async@0.1.1) (2024-08-15)
+
+**Note:** Version bump only for package @rambler-tech/async
+
 # 0.1.0 (2024-08-14)
 
 ### Features