+++ packages/dom/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](https://github.com/rambler-digital-solutions/rambler-common/compare/@rambler-tech/dom@0.2.0...@rambler-tech/dom@0.2.1) (2024-09-27)
+
+**Note:** Version bump only for package @rambler-tech/dom
+
+# [0.2.0](https://github.com/rambler-digital-solutions/rambler-common/compare/@rambler-tech/dom@0.1.0...@rambler-tech/dom@0.2.0) (2024-09-06)
+
+### Features
+
+- **dom:** add attributes to load script ([7610156](https://github.com/rambler-digital-solutions/rambler-common/commit/7610156c85a5dd6e13a61079a8408b264cc4336b))
+
 # 0.1.0 (2024-08-14)
 
 ### Features