+++ packages/lhci-report/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/rambler-digital-solutions/rambler-common/compare/@rambler-tech/lhci-report@0.1.4...@rambler-tech/lhci-report@0.2.0) (2024-10-03)
+
+### Bug Fixes
+
+- **lhci-report:** build to commonjs ([aa64a0d](https://github.com/rambler-digital-solutions/rambler-common/commit/aa64a0d9945323d8f04465760ecd649e4e27e3a3))
+
+### Features
+
+- **lhci-report:** read report dir from lighthouse config ([dcdf8b4](https://github.com/rambler-digital-solutions/rambler-common/commit/dcdf8b4efc279f9c9d04006919ad5a8d286de801))
+
+## [0.1.4](https://github.com/rambler-digital-solutions/rambler-common/compare/@rambler-tech/lhci-report@0.1.3...@rambler-tech/lhci-report@0.1.4) (2024-09-27)
+
+**Note:** Version bump only for package @rambler-tech/lhci-report
+
 ## [0.1.3](https://github.com/rambler-digital-solutions/rambler-common/compare/@rambler-tech/lhci-report@0.1.2...@rambler-tech/lhci-report@0.1.3) (2024-03-25)
 
 ### Bug Fixes