+++ packages/session-storage/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.5](https://github.com/rambler-digital-solutions/rambler-common/compare/@rambler-tech/session-storage@0.1.4...@rambler-tech/session-storage@0.1.5) (2024-09-27)
+
+**Note:** Version bump only for package @rambler-tech/session-storage
+
 ## [0.1.4](https://github.com/rambler-digital-solutions/rambler-common/compare/@rambler-tech/session-storage@0.1.3...@rambler-tech/session-storage@0.1.4) (2024-03-25)
 
 ### Bug Fixes