+++ packages/transport/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/rambler-digital-solutions/rambler-common/compare/@rambler-tech/transport@0.1.1...@rambler-tech/transport@0.2.0) (2024-09-27)
+
+### Features
+
+- **transport:** add request factory ([a2f6e59](https://github.com/rambler-digital-solutions/rambler-common/commit/a2f6e59d5123406eadf37373de6123278553bcd4))
+
+## [0.1.1](https://github.com/rambler-digital-solutions/rambler-common/compare/@rambler-tech/transport@0.1.0...@rambler-tech/transport@0.1.1) (2024-09-27)
+
+**Note:** Version bump only for package @rambler-tech/transport
+
 # 0.1.0 (2024-08-15)
 
 ### Features